# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "XSocial Laravel API",
        "description": "✊✊✊",
        "contact": {
            "email": "ngoctoan06011998@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/topics": {
            "get": {
                "tags": [
                    "Topics"
                ],
                "summary": "topics.index",
                "operationId": "8e16748db53ac966e8af74d772501056",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "users|messages|last_message",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "-id, -created_at, -updated_at",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Topics"
                ],
                "summary": "topics.store",
                "description": "type: DUOS, GROUP",
                "operationId": "fead0ba561f17de0c4c6bccc6f8605e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "recipient_ids",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "recipient_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/topics/{id}": {
            "get": {
                "tags": [
                    "Topics"
                ],
                "summary": "topics.show",
                "operationId": "615c928c6685afe8fd57e2961667b24f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "users|messages|last_message",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "-id, -created_at, -updated_at",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Topics"
                ],
                "summary": "topics.destroy",
                "operationId": "37b43ac2b6d2aed36106fb9c32e65964",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "users.index",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "filter[keyword]",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "roles|followers",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "-id, -created_at",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "users.show",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "filter[keyword]",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "roles|followers",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Bearer {token}"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}